+++ tests/moduletool/common.py
@@ -15,24 +15,17 @@
 
     Contact: code@inmanta.com
 """
-import configparser
 import os
-import shutil
 import subprocess
 import tempfile
-from dataclasses import dataclass
 from subprocess import CalledProcessError
-from typing import Optional, Sequence, Union
+from typing import Optional
 
 import yaml
-from pkg_resources import Requirement
 
-from inmanta import const, module
 from inmanta.config import Config
 from inmanta.module import InstallMode, Project
 from inmanta.moduletool import ModuleTool
-from libpip2pi.commands import dir2pi
-from packaging import version
 
 
 def makeproject(reporoot, name, deps=[], imports=None, install_mode: Optional[InstallMode] = None):
@@ -197,82 +190,3 @@ class BadModProvider(object):
             return getattr(self.parent, method_name)(*args, **kw)
 
         return delegator
-
-
-@dataclass
-class PipIndex:
-    """
-    Local pip index that makes use of dir2pi to publish its artifacts.
-    """
-
-    artifact_dir: str
-
-    @property
-    def url(self) -> str:
-        return f"{self.artifact_dir}/simple"
-
-    def publish(self) -> None:
-        dir2pi(argv=["dir2pi", self.artifact_dir])
-
-
-def module_from_template(
-    source_dir: str,
-    dest_dir: str,
-    *,
-    new_version: Optional[version.Version] = None,
-    new_name: Optional[str] = None,
-    new_requirements: Optional[Sequence[Union[module.InmantaModuleRequirement, Requirement]]] = None,
-    install: bool = False,
-    editable: bool = False,
-    publish_index: Optional[PipIndex] = None,
-    new_content_init_cf: Optional[str] = None,
-) -> module.ModuleV2Metadata:
-    """
-    Creates a v2 module from a template.
-
-    :param source_dir: The directory where the original module lives.
-    :param dest_dir: The directory to use to copy the original to and to stage any changes in.
-    :param new_version: The new version for the module, if any.
-    :param new_name: The new name of the inmanta module, if any.
-    :param new_requirements: The new requirements for the module, if any.
-    :param install: Install the newly created module with the module tool. Requires virtualenv to be installed in the
-        python environment unless editable is True.
-    :param editable: Whether to install the module in editable mode, ignored if install is False.
-    :param publish_index: Publish to the given local path index. Requires virtualenv to be installed in the python environment.
-    :param new_content_init_cf: The new content of the _init.cf file.
-    """
-    # preinstall older version of module
-    shutil.copytree(source_dir, dest_dir)
-    config_file: str = os.path.join(dest_dir, module.ModuleV2.MODULE_FILE)
-    config: configparser.ConfigParser = configparser.ConfigParser()
-    config.read(config_file)
-    if new_version is not None:
-        config["metadata"]["version"] = str(new_version)
-        if new_version.is_devrelease:
-            config["egg_info"] = {"tag_build": f".dev{new_version.dev}"}
-    if new_name is not None:
-        os.rename(
-            os.path.join(
-                dest_dir, const.PLUGINS_PACKAGE, module.ModuleV2Source.get_inmanta_module_name(config["metadata"]["name"])
-            ),
-            os.path.join(dest_dir, const.PLUGINS_PACKAGE, new_name),
-        )
-        config["metadata"]["name"] = module.ModuleV2Source.get_python_package_name(new_name)
-    if new_requirements:
-        config["options"]["install_requires"] = "\n    ".join(
-            str(req if isinstance(req, Requirement) else module.ModuleV2Source.get_python_package_requirement(req))
-            for req in new_requirements
-        )
-    if new_content_init_cf is not None:
-        init_cf_file = os.path.join(dest_dir, "model", "_init.cf")
-        with open(init_cf_file, "w", encoding="utf-8") as fd:
-            fd.write(new_content_init_cf)
-    with open(config_file, "w") as fh:
-        config.write(fh)
-    if install:
-        ModuleTool().install(editable=editable, path=dest_dir)
-    if publish_index is not None:
-        ModuleTool().build(path=dest_dir, output_dir=publish_index.artifact_dir)
-        publish_index.publish()
-    with open(config_file, "r") as fh:
-        return module.ModuleV2Metadata.parse(fh)
